Remove unnecessary print statements
