+++ callflow/modules/auxiliary.py
@@ -11,64 +11,70 @@ import math
 import pandas as pd
 import numpy as np
 
+from .gradients import Gradients
+from .boxplot import BoxPlot
+
 #------------------------------------------------------------------------------
-# CallFlow imports
+# CallFlow imports 
 import callflow
-from callflow.modules.gradients import Gradients
-from callflow.modules.boxplot import BoxPlot
-from callflow.timer import Timer
 
 LOGGER = callflow.get_logger(__name__)
+from callflow.timer import Timer
 
-
-class FastEnsembleAuxiliary:
+#------------------------------------------------------------------------------
+class EnsembleAuxiliary:
     def __init__(
         self,
-        states,
+        gf=callflow.GraphFrame,
+        datasets=[],
+        props={},
         MPIBinCount="20",
         RunBinCount="20",
-        datasets=[],
-        config={},
         process=True,
         write=False,
-        topCallsite=10,
     ):
-        self.timer = Timer()
-        self.df = self.select_rows(states["ensemble_entire"].new_gf.df, datasets)
+        self.gf = gf
         self.MPIBinCount = MPIBinCount
         self.RunBinCount = RunBinCount
-        self.config = config
-        self.states = states
+        self.timer = Timer()
+        self.props = props
+        self.datasets = self.props["dataset_names"]
+
+        self.df = self.select_rows(self.gf.df, self.datasets)
+
         self.process = process
         self.write = write
-        self.datasets = datasets
 
-        self.props = ["rank", "name", "dataset", "all_ranks"]
+        self.hist_props = ["rank", "name", "dataset", "all_ranks"]
         self.filter = True
 
-        """
-        self.runPath = (
-            self.config.callflow_path
-            + "/"
-            + self.config.save_path
-            + "/"
-            + self.config.runName
-        )
-        self.h5IndexFilename = self.runPath + "/h5_index.json"
-        self.moduleh5File = self.runPath + "/module_data.h5"
-        self.callsiteh5File = self.runPath + "/callsite_data.h5"
-        """
-        self.h5IndexFilename = os.path.join(self.config.save_path, "h5_index.json")
-        self.moduleh5File = os.path.join(self.config.save_path, "module_data.h5")
-        self.callsiteh5File = os.path.join(self.config.save_path, "callsite_data.h5")
+        if process:
+            self.compute()
+        else:
+            self.read()
+        LOGGER.info(self.timer)
 
-        self.topCallsite = topCallsite
+    def compute(self):
+        ret = {}
+        path = os.path.join(self.props["save_path"], "ensemble/auxiliary_data.json")
 
-        with self.timer.phase("Group data with indexes"):
+        LOGGER.info("Calculating Gradients, Mean runtime variations, and Distribution.")
+        with self.timer.phase("Process data"):
             self.group_frames()
+        with self.timer.phase("Collect Callsite data"):
+            ret["callsite"] = self.callsite_data()
+        with self.timer.phase("Collect Module data"):
+            ret["module"] = self.module_data()
+        with self.timer.phase("Module callsite map data"):
+            ret["moduleCallsiteMap"] = self.get_module_callsite_map()
+        # with self.timer.phase("Callsite module map data"):
+        #     ret['callsiteModuleMap'] = self.get_callsite_module_map()
+        # if self.write:
+        with self.timer.phase("Writing data"):
+            with open(path, "w") as f:
+                json.dump(ret, f)
 
-        self.callsiteMap = {}
-        self.moduleMap = {}
+        return ret
 
     def filter_dict(self, result):
         ret = {}
@@ -80,14 +86,14 @@ class FastEnsembleAuxiliary:
         ret["callsite"] = {}
 
         group_df = self.df.groupby(["name"]).mean()
-        if self.config.filter_by == "time":
+        if self.props["filter_by"] == "time":
             f_group_df = group_df.loc[
-                group_df[self.config.filter_by] > self.config.filter_below
+                group_df[self.props["filter_by"]] > self.props["filter_below"]
             ]
-        elif self.config.filter_by == "time (inc)":
+        elif self.props["filter_by"] == "time (inc)":
             f_group_df = group_df.loc[
-                group_df[self.config.filter_by]
-                > 0.01 * self.config.filter_perc * group_df["time (inc)"].max()
+                group_df[self.props["filter_by"]]
+                > 0.01 * self.props["filter_perc"] * group_df["time (inc)"].max()
             ]
         callsites = f_group_df.index.values.tolist()
 
@@ -102,13 +108,14 @@ class FastEnsembleAuxiliary:
                     count += 1
         return ret
 
-    def filter_frames(self, nCallsites, attr):
-        xgroup_df = self.df.groupby(["name"]).mean()
-        sort_xgroup_df = xgroup_df.sort_values(by=[attr], ascending=False)
-        nCallsites_df = sort_xgroup_df.nlargest(nCallsites, attr)
-        return nCallsites_df
-
     def group_frames(self):
+        if self.filter:
+            xgroup_df = self.df.groupby(["name"]).mean()
+            sort_xgroup_df = xgroup_df.sort_values(by=["time (inc)"], ascending=False)
+            top100callsites = sort_xgroup_df.nlargest(50, "time (inc)")
+            self.df = self.df[self.df["name"].isin(top100callsites.index.values)]
+
+        self.df.drop(["rank"], axis=1)
         self.module_name_group_df = self.df.groupby(["module", "name"])
         self.module_group_df = self.df.groupby(["module"])
         self.name_group_df = self.df.groupby(["name"])
@@ -128,6 +135,122 @@ class FastEnsembleAuxiliary:
             self.target_name_group_df[dataset] = self.target_df[dataset].groupby(
                 ["name"]
             )
+
+    # Callsite grouped information
+    def callsite_data(self):
+        ret = {}
+
+        # Create the data dict.
+        ensemble = {}
+        for callsite, callsite_df in self.name_group_df:
+            callsite_ensemble_df = self.name_group_df.get_group(callsite)
+            hists = {}
+            hists["Inclusive"] = {}
+            hists["Exclusive"] = {}
+            for prop in self.hist_props:
+                prop_histograms = self.histogram_by_property_ensemble(
+                    callsite_ensemble_df, prop
+                )
+                hists["Inclusive"][prop] = prop_histograms["Inclusive"]
+                hists["Exclusive"][prop] = prop_histograms["Exclusive"]
+
+            gradients = Gradients(self.target_df, binCount=self.RunBinCount).run(
+                columnName="name", callsiteOrModule=callsite
+            )
+            boxplot = BoxPlot(callsite_df)
+            ensemble[callsite] = self.pack_json(
+                callsite_df,
+                callsite,
+                gradients=gradients,
+                q=boxplot.q,
+                outliers=boxplot.outliers,
+                prop_hists=hists,
+            )
+
+        ret["ensemble"] = ensemble
+
+        ## Target data.
+        # Loop through datasets and group the callsite by name.
+        for dataset in self.datasets:
+            name_grouped = self.target_name_group_df[dataset]
+            target = {}
+            for callsite, callsite_df in name_grouped:
+                callsite_ensemble_df = self.name_group_df.get_group(callsite)
+                callsite_target_df = callsite_df
+
+                if not callsite_df.empty:
+                    hists = {}
+                    hists["Inclusive"] = {}
+                    hists["Exclusive"] = {}
+                    for prop in self.hist_props:
+                        prop_histograms = self.histogram_by_property(
+                            callsite_ensemble_df, callsite_target_df, prop
+                        )
+                        hists["Inclusive"][prop] = prop_histograms["Inclusive"]
+                        hists["Exclusive"][prop] = prop_histograms["Exclusive"]
+
+                    boxplot = BoxPlot(callsite_df)
+                    target[callsite] = self.pack_json(
+                        df=callsite_target_df,
+                        name=callsite,
+                        prop_hists=hists,
+                        q=boxplot.q,
+                        outliers=boxplot.outliers,
+                    )
+            ret[dataset] = target
+
+        return ret
+
+    def module_data(self):
+        ret = {}
+        # Module grouped information
+        modules = self.df["module"].unique()
+        ensemble = {}
+        for module, module_df in self.module_group_df:
+            module_ensemble_df = self.module_group_df.get_group(module)
+            hists = {"Inclusive": {}, "Exclusive": {}}
+            for prop in self.hist_props:
+                prop_histograms = self.histogram_by_property_ensemble(
+                    module_ensemble_df, prop
+                )
+                hists["Inclusive"][prop] = prop_histograms["Inclusive"]
+                hists["Exclusive"][prop] = prop_histograms["Exclusive"]
+
+            # Calculate gradients
+            gradients = Gradients(self.target_df, binCount=self.RunBinCount).run(
+                columnName="module", callsiteOrModule=module
+            )
+            ensemble[module] = self.pack_json(
+                df=module_df, name=module, gradients=gradients, prop_hists=hists
+            )
+
+        ret["ensemble"] = ensemble
+
+        for dataset in self.datasets:
+            target = {}
+            module_group_df = self.target_module_group_df[dataset]
+            for module, module_df in module_group_df:
+                module_ensemble_df = self.module_group_df.get_group(module)
+                module_target_df = module_df
+                gradients = {"Inclusive": {}, "Exclusive": {}}
+                hists = {"Inclusive": {}, "Exclusive": {}}
+                if not module_target_df.empty:
+                    for prop in self.hist_props:
+                        prop_histograms = self.histogram_by_property(
+                            module_ensemble_df, module_target_df, prop
+                        )
+                        hists["Inclusive"][prop] = prop_histograms["Inclusive"]
+                        hists["Exclusive"][prop] = prop_histograms["Exclusive"]
+                    target[module] = self.pack_json(
+                        df=module_target_df,
+                        name=module,
+                        gradients=gradients,
+                        prop_hists=hists,
+                    )
+
+            ret[dataset] = target
+
+        return ret
 
     def select_rows(self, df, search_strings):
         unq, IDs = np.unique(df["dataset"], return_inverse=True)
@@ -205,8 +328,8 @@ class FastEnsembleAuxiliary:
             "id": "node-" + str(df["nid"].tolist()[0]),
             "dataset": df["dataset"].unique().tolist(),
             "module": df["module"].tolist()[0],
-            "callers": df["callers"].unique().tolist(),
-            "callees": df["callees"].unique().tolist(),
+            # "callers": df["callers"].unique().tolist(),
+            # "callees": df["callees"].unique().tolist(),
             "component_path": df["component_path"].unique().tolist(),
             "component_level": df["component_level"].unique().tolist(),
             "Inclusive": {
@@ -260,6 +383,7 @@ class FastEnsembleAuxiliary:
             time_ensemble_exclusive_arr = np.array(ensemble_df["time"].tolist())
 
         elif prop == "rank":
+            ensemble_df.reset_index(drop=True, inplace=True)
             ensemble_prop = ensemble_df.groupby(["dataset", prop])[
                 ["time", "time (inc)"]
             ].mean()
@@ -304,9 +428,12 @@ class FastEnsembleAuxiliary:
             time_target_inclusive_arr = np.array(target_df["time (inc)"].tolist())
             time_target_exclusive_arr = np.array(target_df["time"].tolist())
         elif prop == "rank":
+            ensemble_df.reset_index(drop=True, inplace=True)
             ensemble_prop = ensemble_df.groupby(["dataset", prop])[
                 ["time", "time (inc)"]
             ].mean()
+
+            target_df.reset_index(drop=True, inplace=True)
             target_prop = target_df.groupby(["dataset", prop])[
                 ["time", "time (inc)"]
             ].mean()
@@ -360,237 +487,4 @@ class FastEnsembleAuxiliary:
             "ensemble": self.histogram_format(histogram_ensemble_exclusive_grid),
             "target": self.histogram_format(histogram_target_exclusive_grid),
         }
-        return ret
-
-    # Callsite grouped information
-    def callsite_data_old(self):
-        ret = {}
-
-        # Create the data dict.
-        ensemble = {}
-        for callsite, callsite_df in self.name_group_df:
-            callsite_ensemble_df = self.name_group_df.get_group(callsite)
-            hists = {}
-            hists["Inclusive"] = {}
-            hists["Exclusive"] = {}
-            for prop in self.props:
-                prop_histograms = self.histogram_by_property_ensemble(
-                    callsite_ensemble_df, prop
-                )
-                hists["Inclusive"][prop] = prop_histograms["Inclusive"]
-                hists["Exclusive"][prop] = prop_histograms["Exclusive"]
-
-            gradients = KDE_gradients(self.target_df, binCount=self.RunBinCount).run(
-                columnName="name", callsiteOrModule=callsite
-            )
-            boxplot = BoxPlot(callsite_df)
-            ensemble[callsite] = self.pack_json(
-                callsite_df,
-                callsite,
-                gradients=gradients,
-                q=boxplot.q,
-                outliers=boxplot.outliers,
-                prop_hists=hists,
-            )
-
-        ret["ensemble"] = ensemble
-
-        ## Target data.
-        # Loop through datasets and group the callsite by name.
-        for dataset in self.datasets:
-            name_grouped = self.target_name_group_df[dataset]
-            target = {}
-            for callsite, callsite_df in name_grouped:
-                callsite_ensemble_df = self.name_group_df.get_group(callsite)
-                callsite_target_df = callsite_df
-
-                if not callsite_df.empty:
-                    hists = {}
-                    hists["Inclusive"] = {}
-                    hists["Exclusive"] = {}
-                    for prop in self.props:
-                        prop_histograms = self.histogram_by_property(
-                            callsite_ensemble_df, callsite_target_df, prop
-                        )
-                        hists["Inclusive"][prop] = prop_histograms["Inclusive"]
-                        hists["Exclusive"][prop] = prop_histograms["Exclusive"]
-
-                    boxplot = BoxPlot(callsite_df)
-                    target[callsite] = self.pack_json(
-                        df=callsite_target_df,
-                        name=callsite,
-                        prop_hists=hists,
-                        q=boxplot.q,
-                        outliers=boxplot.outliers,
-                    )
-            ret[dataset] = target
-
-    #     return ret
-
-    def get_data_from_hd5(self, nodes, col):
-        ret = {}
-        if col == "module":
-            filename = self.moduleh5File
-            mapping = self.moduleMap
-
-        elif col == "name":
-            filename = self.callsiteh5File
-            mapping = self.callsiteMap
-
-        ensemble = {}
-        for node in nodes:
-            module_ensemble_df = pd.read_hdf(filename, key=mapping[node])
-            hists = {"Inclusive": {}, "Exclusive": {}}
-            for prop in self.props:
-                prop_histograms = self.histogram_by_property_ensemble(
-                    module_ensemble_df, prop
-                )
-                hists["Inclusive"][prop] = prop_histograms["Inclusive"]
-                hists["Exclusive"][prop] = prop_histograms["Exclusive"]
-
-            # Calculate gradients
-            gradients = {"Inclusive": {}, "Exclusive": {}}
-            gradients = KDE_gradients(self.target_df, binCount=self.RunBinCount).run(
-                columnName=col, callsiteOrModule=node
-            )
-            quartiles = {"Inclusive": {}, "Exclusive": {}}
-            outliers = {"Inclusive": {}, "Exclusive": {}}
-            if col == "name":
-                boxplot = BoxPlot(module_ensemble_df)
-                quartiles = boxplot.q
-                outliers = boxplot.outliers
-
-            ensemble[node] = self.pack_json(
-                df=module_ensemble_df,
-                name=node,
-                gradients=gradients,
-                prop_hists=hists,
-                q=quartiles,
-                outliers=outliers,
-            )
-
-        ret["ensemble"] = ensemble
-
-        for dataset in self.datasets:
-            target = {}
-
-            module_target_df = module_ensemble_df.loc[
-                module_ensemble_df["dataset"] == dataset
-            ]
-            for node in nodes:
-                gradients = {"Inclusive": {}, "Exclusive": {}}
-                hists = {"Inclusive": {}, "Exclusive": {}}
-                quartiles = {"Inclusive": {}, "Exclusive": {}}
-                outliers = {"Inclusive": {}, "Exclusive": {}}
-
-                if module_target_df.shape[0] != 0:
-                    for prop in self.props:
-                        prop_histograms = self.histogram_by_property(
-                            module_ensemble_df, module_target_df, prop
-                        )
-                        hists["Inclusive"][prop] = prop_histograms["Inclusive"]
-                        hists["Exclusive"][prop] = prop_histograms["Exclusive"]
-
-                    if col == "name":
-                        boxplot = BoxPlot(module_target_df)
-                        quartiles = boxplot.q
-                        outliers = boxplot.outliers
-
-                    target[node] = self.pack_json(
-                        df=module_target_df,
-                        name=node,
-                        gradients=gradients,
-                        prop_hists=hists,
-                        q=quartiles,
-                        outliers=outliers,
-                    )
-
-            ret[dataset] = target
-
-        return ret
-
-    def module_data(self):
-        ret = {}
-        module_group_df = self.df.groupby(["module"])
-
-        self.moduleMap = {}
-        count = 0
-        for module, module_df in module_group_df:
-            module_ensemble_df = module_group_df.get_group(module)
-            key = "module_" + str(count)
-            self.moduleMap[module] = key
-            module_ensemble_df.to_hdf(self.moduleh5File, key=key)
-            count += 1
-
-    def callsite_data(self):
-        ret = {}
-        name_group_df = self.df.groupby(["name"])
-        count = 0
-        for name, name_df in name_group_df:
-            callsite_ensemble_df = name_group_df.get_group(name)
-            key = "callsite_" + str(callsite_ensemble_df["nid"].unique()[0])
-            self.callsiteMap[name] = key
-            callsite_ensemble_df.to_hdf(self.callsiteh5File, key=key)
-            count += 1
-
-    def write_maps(self):
-        # as requested in comment
-        exDict = {"moduleMap": self.moduleMap, "callsiteMap": self.callsiteMap}
-
-        with open(self.h5IndexFilename, "w") as file:
-            print(exDict)
-            file.write(json.dumps(exDict))
-            print(f"writen to file : {self.h5IndexFilename}")
-
-    def read_maps(self):
-        f = open(self.h5IndexFilename, "r")
-        data = json.load(f)
-
-        self.moduleMap = data["moduleMap"]
-        self.callsiteMap = data["callsiteMap"]
-
-        print("Read the h5 index map.")
-
-    def run(self):
-        print("Calculating Gradients, Mean runtime variations, and Distribution.")
-        with self.timer.phase("Collect Callsite data"):
-            self.callsite_data()
-        with self.timer.phase("Collect Module data"):
-            self.module_data()
-        with self.timer.phase("Write module's and callsite's hdf indexes"):
-            self.write_maps()
-
-        print(self.timer)
-
-    def fetch(self):
-        ret = {}
-
-        with self.timer.phase("Read module's and callsite's hdf indexes"):
-            self.read_maps()
-
-        with self.timer.phase("Collect Callsite data"):
-            modules = self.df["module"].unique().tolist()
-
-        with self.timer.phase("Filter"):
-            if self.filter:
-                # topCallsites_df = self.filter_frames(self.topCallsite, "time (inc)")
-                topCallsites_df = self.filter_frames(self.topCallsite, "time")
-                callsites = topCallsites_df.index.values
-                self.df = self.df[self.df["name"].isin(topCallsites_df.index.values)]
-
-            else:
-                callsites = self.df["name"].unique().tolist()
-
-        with self.timer.phase("Fetch module"):
-            ret["module"] = {}
-            ret["module"] = self.get_data_from_hd5(modules, "module")
-
-        with self.timer.phase("Fetch callsite"):
-            ret["callsite"] = self.get_data_from_hd5(callsites, "name")
-
-        with self.timer.phase("Module callsite map data"):
-            ret["moduleCallsiteMap"] = self.get_module_callsite_map()
-
-        print(self.timer)
-
         return ret
